livecheck: Add support for POST requests
livecheck currently doesn't support `POST` requests but it wasn't
entirely clear how best to handle that. I initially approached it as
a `Post` strategy but unfortunately that would have required us to
handle response body parsing (e.g., JSON, XML, etc.) in some fashion.
We could borrow some of the logic from related strategies but we would
still be stuck having to update `Post` whenever we add a strategy for
a new format.

Instead, this implements `POST` support by borrowing ideas from the
`using: :post` and `data` `url` options found in formulae. The notable
differences are that this uses `method: :post` (instead of `using`)
and this also supports a `json` option (in place of `data`), which
will pass `--json <value>` to curl instead of `--data <value>`.

With this approach, we can make a `POST` request using any strategy
that calls `Strategy::page_headers` or `::page_content` (directly or
indirectly) and everything else works the same as usual. The only
change needed in related strategies was to pass the options through
to the `Strategy` methods.

That is to say, if we need to parse a JSON response from a `POST`
request, we add `method: :post` and a `data`/`json` hash or string to
the `livecheck` block `url` and use `strategy :json` (with a
`strategy` block). This leans on existing patterns that we're already
familiar with and shouldn't require any notable maintenance burden
when adding new strategies, so it seems like a better approach than a
`Post` strategy.